"Refactor FuelTankBlock constructor for clarity"
This commit simplifies the method for constructing FuelTankBlocks in the application. It removes unnecessary static methods in the FuelTankBlock classes in both the forge and fabric directories, changing the visibility of the FuelTankBlock constructor to 'public' and direct using new operator in "CRBlocksImpl.java" files under both directories. Additionally, this commit streamlines the 'PORTABLE_FUEL_INTERFACE' block entity entry in the "CRBlockEntitiesImpl.java" file by merging the variable assignment with the block entity registration call. This refactor aims to enhance code readability and maintainability by reducing superfluous code and improving directness.